cleaned up and simplified
